% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_phy.R
\name{cor_phy}
\alias{cor_phy}
\title{Correlates each lipid with all OTUs}
\usage{
cor_phy(phy, lipids, method = c("pearson", "spearman", "kendall"),
  level = "Phylum", display = "all", shape = 19, shape.other = 1,
  alpha = 0.75, alpha.other = 0.2, col.other = "gray75",
  colors = c("random", "distinc"), luminosity = c("random", "light",
  "bright", "dark"), flip = TRUE, box.plot = FALSE, width = 0.25,
  aspect = 1)
}
\arguments{
\item{phy}{Phyloseq object containing an OTU and taxonomy table.}

\item{lipids}{Data frame with concentration of lipids (or any other variables/parameters) to correlate with phylogenetic data. Rows are observations/samples and must match the samples in 'phy'. Used by 'phlo_lipids'}

\item{method}{correaltion coefficient used in 'cor' {stats}, i.e., "pearson", "kendall", or "spearman"}

\item{level}{taxonomic level to evaluate, if desplay is set to "all"}

\item{display}{taxa to be displayed/evaluated. The remining taxa will be subsumed as '_other'. If 'display = "all"',
all taxa of the specified 'level' will be displayed (can be slow for lower taxonomic levels).}

\item{shape}{plot shape of 'display' taxa}

\item{shape.other}{plot shape of '_other' taxa}

\item{alpha}{opacitiy of 'display' taxa}

\item{alpha.other}{opacity of '_other' taxa}

\item{col.other}{color of '_other' taxa}

\item{colors}{type of color palette from 'randomcolorR::randomColor',}

\item{luminosity}{luminosity setting of solor palette, if 'colors = "random"'}

\item{flip}{flips axis; if TRUE, horizontal plot}

\item{box.plot}{if TRUE, a boxplot will be drawn instead of a jitter plot}

\item{width}{width of jitter (or boxplot)}

\item{aspect}{aspect ratio of plot (direction depends of 'flip')}
}
\value{
A list with two data frames containing summaries (max and average coefficients) by taxa (.$taxa) and lipids (.$lipids), respectively, as well as a jitter plot (.$plot).
}
\description{
This function allows you to extract the phylogeneic affiliations of OTUs, using a theshold correlation coefficient
Returns a table with each row representing one OTU, and each column one lipid compound. Values are affiliations of above-threshold OTUs.
'lower' = below threshold, 'other' = sumsumed according to cutoff fraction
}
\seealso{
Other Main functions: \code{\link{cor_tax}},
  \code{\link{norm_dna}}, \code{\link{phylo_lipids}}
}
\concept{Main functions}
