% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_tax.R
\name{cutoff_tax}
\alias{cutoff_tax}
\title{applies cutoff abundance}
\usage{
cutoff_tax(phy, cutoff = 0.03, level = "Phylum", exempt = FALSE)
}
\arguments{
\item{phy}{Phyloseq object containing at least an OTU and a taxonomy table.}

\item{cutoff}{cutoff abundance for assignment of 'other'}

\item{level}{taxonomic level}

\item{exempt}{character vector of taxa exempt from cutoff}
}
\value{
A list with i elements, containing the filtered OTU/taxa names in the i-th group
}
\description{
cutoff for subsummation as 'other, expemtions to this rule
}
\keyword{internal}
