% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_proteo.R
\name{diff_proteo}
\alias{diff_proteo}
\title{Replaces the phylum Proteobacteria with the classes (i.e., Aplpha-, Beta-, Deltaproteobacteria, etc.)}
\usage{
diff_proteo(phy)
}
\arguments{
\item{phy}{Phyloseq object containing an OTU and taxonomy table.}
}
\description{
Aussumes that second taxonomic rank is 'Phylum'
}
\examples{

}
