% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_dna.R
\name{norm_dna}
\alias{norm_dna}
\title{normalizes read numbers}
\usage{
norm_dna(phy, dna = NULL)
}
\arguments{
\item{phy}{Phyloseq object containing an OTU and taxonomy table.}

\item{dna}{numeric vector containing absolute DNA abundance in µg,
which will be multipled with the relative OTU abudances. If NULL (default), OTU abundances will be normalized to the sum of all OUTs, i.e. relative abundanes}
}
\description{
Returns either relative abundances (relative == TRUE), or OTU-specific DNA abundances in [ng] (!)
}
\seealso{
Other Main functions: \code{\link{cor_phy}},
  \code{\link{cor_tax}}, \code{\link{phylo_lipids}}
}
\concept{Main functions}
\keyword{DNA}
\keyword{OTU-specific}
\keyword{concentrations}
\keyword{normalize}
