% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.lipids.R
\docType{package}
\name{phylo.lipids-package}
\alias{phylo.lipids}
\alias{phylo.lipids-package}
\title{phylo.lipids: Functions to compare environmental lipidomic and meta taxonomic data}
\description{
Allows you to explore and visualize environmental meta-taxonomic data, 
   and to correlate it with other environmental variables 
   (e.g. concentration of lipid biomarkers).
}
\references{
Yuki Weber, Jaap S. Sinninghe Damsté, Jakob Zopfi, Cindy De Jonge, Adrian Gili, Carsten J. Schubert, Fabio Lepori, Moritz F. Lehmann, Helge Niemann (2018). "Redox-dependent niche differentiation of tetraether lipid producing bacteria: Evidence for multiple branched GDGT sources in lakes"
}
\keyword{internal}
