% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_a_group.R
\name{plot_a_group}
\alias{plot_a_group}
\title{helper function: plots standardized otu abundances}
\usage{
plot_a_group(phy, taxnames, label = "NULL", otus = TRUE,
  ribbon = TRUE, ncol = 5, is.numeric.label = F,
  standardize = c("sum", "max"), wt.means = F, rev.scale = "NULL")
}
\arguments{
\item{phy}{Phyloseq object}

\item{taxnames}{A (list of) char vector(s) containing otu names to be plotted
If list, output will be facetted by the elements of the list.}

\item{label}{a element of phy'at'sam_data to replace the original sample names}

\item{otus}{display individual otus?}

\item{ribbon}{Plot error bounds (+/- 2*SD)}

\item{ncol}{Number of columns in facets}

\item{is.numeric.label}{If TRUE, force coertion of labels to numeric}

\item{standardize}{Method of standardization of the otu abundances,
either to the sum or the maximum across all samples}

\item{wt.means}{Plot weighted means by abundance of the otus (!!! under development !!!)}

\item{rev.scale}{force reverse scale ON/OFF (reverse is otherwise automatic for label =='depth')}

\item{x}{character vector with otu names}
}
\value{
ggplot line plot
}
\description{
helper function: plots standardized otu abundances
}
