% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_groups.R
\name{plot_groups}
\alias{plot_groups}
\title{Plots vertical profiles of OTU groups with taxonomy summary}
\usage{
plot_groups(phy, level = "Phylum", label = "NULL", k = 10,
  cutoff = 0.03, exempt = FALSE, otus = T, ribbon = T,
  prune_groups = NULL, count = c("n", "sum"), fancy = T,
  highlight = "NULL", standardize = c("sum", "max"), ncol = 5,
  key_col = 2, clust = c("ward.D2", "ward.D", "single", "complete",
  "average", "mcquitty", "median", "centroid"))
}
\arguments{
\item{phy}{Phyloseq object containing at least an OTU and a taxonomy table.}

\item{level}{taxonomic level}

\item{label}{A column of 'phy'at'samdata' that is used to label observations (character string),
or a vector of same lengh as observations}

\item{cutoff}{cutoff abundance for assignment of 'other'}

\item{exempt}{character vector of taxa exempt from cutoff}

\item{ncol}{Number of columns in the arranged plots/facets}

\item{key_col}{Number of columns of the color key}

\item{clust}{Hclust method passed to hclust() in group_tax()}
}
\value{
a summary plot
}
\description{
Plots vertical profiles of OTU groups with taxonomy summary
}
