% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pie.R
\name{plot_pie}
\alias{plot_pie}
\title{helper function: Plots taxonomy pie chart}
\usage{
plot_pie(x, count = c("n", "avg"))
}
\arguments{
\item{x}{data frame or tibble containing the columns:
1) Taxonomy (character)
2) number of taxa (integers)
3) average abundance}
}
\value{
ggplot pie chart
}
\description{
helper function: Plots taxonomy pie chart
}
\keyword{internal}
