% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_tax.R
\name{standardize_tax}
\alias{standardize_tax}
\title{Standardized otu abundances}
\usage{
standardize_tax(phy, method)
}
\arguments{
\item{phy}{Phyloseq object containing at least an OTU and a taxonomy table.}

\item{method}{Method of standardization of the otu abundances,
either to the sum or the maximum across all samples}
}
\value{
Returns a data standardized otu table with a 'sum' column containg the sums of the original data
}
\description{
Standardized otu abundances
}
