% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_tax.R
\name{sum_tax}
\alias{sum_tax}
\title{Summarizes taxonomy at a given taxonomic level}
\usage{
sum_tax(phy, cutoff = 0.03, level = "Phylum", exempt = FALSE)
}
\arguments{
\item{phy}{Phyloseq object containing at least an OTU and a taxonomy table.}
}
\value{
A list with i elements, containing the OTU/taxa names in the i-th group
}
\description{
a cutoff can be fifined for subsummation as 'other, as well as expemtions to this rule
}
