% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_taxa_sum.R
\name{trim_taxa_sum}
\alias{trim_taxa_sum}
\title{trim Phyloseq onject by threshold sum across all samples}
\usage{
trim_taxa_sum(phy, min)
}
\arguments{
\item{phy}{Phyloseq object containing an OTU and taxonomy table.}

\item{min}{Threshold value (minimum) for taxa to be returned.
Context dependent: reads, concentrations, relative abundanece, etc}
}
\description{
trim Phyloseq onject by threshold sum across all samples
}
\examples{

}
