% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_groups.R
\docType{data}
\name{x}
\alias{x}
\title{Plots vertical profiles of OTU groups with taxonomy summary}
\format{An object of class \code{list} of length 20.}
\usage{
x
}
\arguments{
\item{phy}{Phyloseq object containing at least an OTU and a taxonomy table.}

\item{x}{list of n groups containing otu names of each group}

\item{label}{A column of 'phy'at'samdata' that is used to label observations (character string), or a vector of same lengh as observations}
}
\value{
a summary plot
}
\description{
Plots vertical profiles of OTU groups with taxonomy summary
}
\keyword{datasets}
