% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tax.R
\name{group_tax}
\alias{group_tax}
\title{Finds groups of OTUs/taxa with similar spatial distribution}
\usage{
group_tax(phy, cut = F, sq = F, h = FALSE, k = 20,
  dist = c("euclidean", "bray", "..."), clust = c("ward.D2", "ward.D",
  "single", "complete", "average", "mcquitty", "median", "centroid"))
}
\arguments{
\item{phy}{Phyloseq object containing at least an OTU and a taxonomy table.}
}
\value{
A list with i elements, containing the OTU/taxa names in the i-th group
}
\description{
Aussumes that second taxonomic rank is 'Phylum'
}
