% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_a_group.R
\name{plot_a_group}
\alias{plot_a_group}
\title{helper function: plots standardized}
\usage{
plot_a_group(x, phy, label = "NULL")
}
\arguments{
\item{x}{character vector with otu names}

\item{phy}{Phyloseq object}
}
\value{
ggplot gragh
}
\description{
helper function: plots standardized
}
\keyword{internal}
