% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_a_group.R
\name{plot_a_group}
\alias{plot_a_group}
\title{helper function: plots standardized otu abundances}
\usage{
plot_a_group(phy, taxnames, label = "NULL", otus = TRUE,
  ribbon = TRUE, ncol = 5, is.numeric.label = F)
}
\arguments{
\item{phy}{Phyloseq object}

\item{taxnames}{A (list of) char vector(s) containing otu names to be plotted
If list, output will be facetted by the elements of the list.}

\item{label}{a element of phy'at'sam_data to replace the original sample names}

\item{otus}{display individual otus?}

\item{ribbon}{Plot error bounds (+/- 2*SD)}

\item{ncol}{Number of columns in facets}

\item{is.numeric.label}{If TRUE, force coertion of labels to numeric}

\item{x}{character vector with otu names}
}
\value{
ggplot line plot
}
\description{
helper function: plots standardized otu abundances
}
