% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_groups.R
\name{plot_groups}
\alias{plot_groups}
\title{Plots vertical profiles of OTU groups with taxonomy summary}
\usage{
plot_groups(phy, level = "Phylum", label = "NULL",
  is.numeric.label = "NULL", rev.scale = "NULL", use_groups = F,
  k = 10, h = F, cutoff = 0.03, exempt = FALSE, otus = T,
  ribbon = T, prune_groups = NULL, count = c("n", "sum"),
  fancy = T, highlight = "NULL", standardize = c("sum", "max"),
  ncol = NULL, key_col = 2, clust = c("ward.D2", "ward.D", "single",
  "complete", "average", "mcquitty", "median", "centroid"),
  dist = c("euclidean", "bray", "..."))
}
\arguments{
\item{phy}{Phyloseq object containing at least an OTU and a taxonomy table.}

\item{level}{taxonomic level}

\item{label}{A column of 'phy'at'samdata' that is used to label observations (character string),
or a vector of same lengh as observations}

\item{rev.scale}{reverse y scale. automatic if label == "depth"}

\item{use_groups}{Use a previously generated list of otu groups}

\item{cutoff}{cutoff abundance for assignment of 'other'}

\item{exempt}{character vector of taxa exempt from cutoff}

\item{prune_groups}{A character vector with the grop names to be displayed. If NULL, all are shown.}

\item{count}{Determines whether the statistics are based on the NUMBER of otus or their ABUNDANCE SUMS}

\item{highlight}{Taxa to be highlighted}

\item{standardize}{Method of standardization of the otu abundances,}

\item{ncol}{Number of columns in the arranged plots/facets}

\item{key_col}{Number of columns of the color key}

\item{clust}{Hclust method passed to hclust() in group_tax()}

\item{dist}{Distance matrix type passed to 'dist()' to be used in clustering
either to the sum or the maximum across all samples}
}
\value{
A list of summary plots and data
}
\description{
Plots vertical profiles of OTU groups with taxonomy summary
}
